+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This project uses [Semantic Versioning 2.0.0](http://semver.org/), the format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+- Added `initiatePushWithIdentifier` to initiate domain pushes using an account identifier.
+- Added `getName` to `Account`.
+
+### Deprecated
+
+- Deprecated `initiatePush`. Use `initiatePushWithIdentifier` instead.
+
 ## 5.2.0 - 2026-03-23
 
 ### Added